climbing: refactor - extract InteractivePath
